refactor - move default options
