+++ src/components/piechart/options/defaultOptions.js
